refactor: Extracts authenticators
